updates to sankey models
