--- src/app/shared/fsat/sankey.model.ts
@@ -1,12 +1,0 @@
-export interface FsatSankeyNode {
-    name: string,
-    value: number,
-    loss?: number,
-    source?: number,
-    target?: number[],
-    x: number,
-    y: number,
-    nodeColor: string,
-    id?: string,
-    isConnector: boolean
-}
