--- src/app/shared/models/phast/sankey.model.ts
@@ -1,13 +1,0 @@
-export interface PHASTSankeyNode {
-    name: string,
-    value: number,
-    x: number,
-    y: number,
-    nodeColor: string,
-    source: number,
-    target: number[],
-    isConnector: boolean,
-    isConnectingPath?: boolean,
-    isCircularFlow?: boolean,
-    id?: string,
-}
