+++ src/app/shared/models/sankey.ts
@@ -1,12 +1,14 @@
-export interface PsatSankeyNode {
+export interface SankeyNode {
     name: string,
     value: number,
+    loss?: number,
     x: number,
     y: number,
     nodeColor: string,
-    loss: number,
     source: number,
     target: number[],
     isConnector: boolean,
+    isConnectingPath?: boolean,
+    isCircularFlow?: boolean,
     id?: string,
 }
