--- src/app/shared/models/steam/sankey.model.ts
@@ -1,13 +1,0 @@
-export interface SSMTSankeyNode {
-    name: string,
-    value: number,
-    x: number,
-    y: number,
-    nodeColor: string,
-    source: number,
-    target: number[],   
-    isConnector: boolean,
-    isConnectingPath?: boolean,
-    isCircularFlow?: boolean,
-    id?: string,
-}
