--- src/app/shared/pie-chart/pie-chart.component.css
@@ -1,20 +1,0 @@
-.pie-chart-container {
-    width: 100%;
-}
-
-path {
-    fill: none;
-}
-
-
-.polyline {
-    opacity: 0.3;
-    stroke: #000;
-    stroke-width: 2px;
-    fill: none;
-}
-
-.arc {
-    stroke: #aaa;
-    stroke-width: 1px;
-}
