--- src/app/shared/pie-chart/pie-chart.component.html
@@ -1,12 +1,0 @@
-<div class="pie-chart-container chart-container">
-  <div #ngChart class="pie-chart">
-  </div>
-</div>
-
-<div class="col-md-12 hide-print">
-  <div class="graph-btn-container">
-    <span class="fa fa-download" (click)="downloadChart()" (mouseenter)="initTooltip('btnExportChart')" (mouseleave)="hideTooltip('btnExportChart')"></span>
-    <app-simple-tooltip [message]="'Download graph as image'" [hover]="hoverBtnExport" [display]="displayExportTooltip" [adjustLeft]="20"
-      [adjustPercent]="false"></app-simple-tooltip>
-  </div>
-</div>
