--- src/app/shared/pie-chart/pie-chart.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PieChartComponent } from './pie-chart.component';
-
-describe('PieChartComponent', () => {
-  let component: PieChartComponent;
-  let fixture: ComponentFixture<PieChartComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PieChartComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PieChartComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
