+++ src/app/shared/pie-chart/pie-chart.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { PieChartComponent } from './pie-chart.component';
 import { SimpleTooltipModule } from '../simple-tooltip/simple-tooltip.module';
 import { PlotlyPieChartComponent } from '../plotly-pie-chart/plotly-pie-chart.component';
 import { PlotlyBarChartComponent } from '../plotly-bar-chart/plotly-bar-chart.component';
@@ -9,7 +8,6 @@ import { PlotlyBarChartComponent } from '../plotly-bar-chart/plotly-bar-chart.co
 
 @NgModule({
   declarations: [
-    PieChartComponent,
     PlotlyPieChartComponent,
     PlotlyBarChartComponent
   ],
@@ -18,7 +16,6 @@ import { PlotlyBarChartComponent } from '../plotly-bar-chart/plotly-bar-chart.co
     SimpleTooltipModule
   ],
   exports: [
-    PieChartComponent,
     PlotlyPieChartComponent,
     PlotlyBarChartComponent
   ]
