--- src/app/shared/waterfall-graph/waterfall-graph.component.css
@@ -1,7 +1,0 @@
-/* .waterfall-node-label-title {
-    font-weight: bold;
-} */
-
-.waterfall-line {
-    stroke-dasharray: 5,5;
-}
