--- src/app/shared/waterfall-graph/waterfall-graph.component.html
@@ -1,10 +1,0 @@
-<div #ngChartContainer class="panel-chart-container">
-
-  <div class="d-flex w-100 h-100 justify-content-center align-items-center">
-    <div class="row m-0">
-      <div class="col-12 p-0">
-        <div #ngChart></div>
-      </div>
-    </div>
-  </div>
-</div>
